added comments, fixed issue with event __str__
