+++ papas/pfalgo/pfreconstructor.py
@@ -1,122 +1,220 @@
+from heppy.papas.data.identifier import Identifier
+from heppy.papas.graphtools.DAG import Node
+from heppy.papas.pfalgo.pfblocksplitter import BlockSplitter
 from heppy.papas.pdt import particle_data
 from heppy.papas.path import StraightLine, Helix
-from heppy.papas.pfobjects import Particle
-from heppy.papas.data.identifier import Identifier
-
+from heppy.papas.pfobjects import Reconstructed_Particle
 
 from ROOT import TVector3, TLorentzVector
 import math
 import pprint
 
-class PFReconstructor(object):
+#Discuss with colin self.locked vs ecal.locked
+#209 in reconstruct_block extra ecals to be added in
+#remove sort
 
-    def __init__(self, links, detector, logger):
-        self.links = links
+
+class PFReconstructor(object):
+    ''' The reconstructor takes an event containing blocks of elements
+        and attempts to reconstruct particles
+        The following strategy is used (to be checked with Colin)
+        single elements:
+             track -> charged hadron
+             hcal  -> neutral hadron
+             ecal  -> photon
+        connected elements:
+              has more than one hcal
+                -> each hcal is treated using rules below
+              has an hcal with one or more connected tracks
+                -> add up all connected track energies, turn each track into a charged hadron
+                -> add up all ecal energies connected to the above tracks
+                -> if excess = hcal energy + ecal energies - track energies > 0
+                       and excess < ecal energies
+                           then turn the excess into an photon
+                -> if excess > 0 and excess > ecal energies
+                          make a neutral hadron with excess- ecal energies
+                          make photon with ecal energies
+              has hcal but no track (nb by design there will be no attached ecals because hcal ecal links have been removed
+                    so this will equate to single hcal:- that two hcals should not occur as a single block
+                    because if they are close enough to be linked then they should already have been merged)
+                -> make a neutral hadron 
+              
+              has track(s) 
+                -> each track is turned into a charged hadron
+              has track(s) and  ecal(s)
+                -> the tracks are turned into charged hadrons, the ecals are marked as locked but energy is not checked 
+                and no photons are made
+                TODO handle case where there is more energy in ecals than in the track and make some photons
+              has only ecals 
+                -> this should not occur because ecals that are close enough to be linked should already have been merged
+        
+             
+         If history_nodes are provided then the particles are linked into the exisiting history
+         
+         Contains:
+            blocks: the dictionary of blocks to be reconstructed { blockid; block }
+            unused: list of unused elements
+            particles: list of constructed particles
+            history_nodes: optional, desribes links between elements, blocks, particles
+         Example usage:
+         
+              reconstructed = PFReconstructor(event)
+              event.reconstructed_particles= sorted( reconstructed.particles,
+                            key = lambda ptc: ptc.e(), reverse=True)
+              event.history_nodes=reconstructed.history_nodes
+        ''' 
+    
+    def __init__(self,  detector, logger):
         self.detector = detector
         self.log = logger
-        self.reconstruct(links)
-        
-        
-    def reconstruct(self, links):
+    #self.reconstruct(links)
+
+
+    def reconstruct(self, event,  blocksname, historyname):
+        '''arguments event: should contain blocks and optionally history_nodes'''
+        self.blocks = getattr(event,  blocksname)
         self.unused = []
-        self.particles = []
-        all_subgroups = dict()
-        # pprint.pprint( links.groups )
-        # import pdb; pdb.set_trace()
+        self.particles = dict()
         
-        for groupid, group in links.groups.iteritems():
-            if self.simplify_group(group):
-                all_subgroups[groupid] = links.subgroups(groupid)
-        for group_id, subgroups in all_subgroups.iteritems():
-            del links.groups[group_id]
-            links.groups.update(subgroups)
         
-        #ordering of groups to help with matching reconstruction methods        
-        groups =  sorted(links.groups.iteritems(), key = lambda g: len(g), reverse=True )           
+        # history nodes will be used to connect reconstructed particles into the history
+        # its optional at the moment
+        if hasattr(event, historyname):
+            self.history_nodes = event.history_nodes
+        else : 
+            self.history_nodes = None
         
-        for group_id, group in groups:
-            self.log.info( "group {group_id} {group}".format(
-                group_id = group_id,
-                group = group) )
-            #if len(group)< 6: #ALICE debugging  means we only process bigger groups
-            #    continue  
-            self.particles.extend( self.reconstruct_group(group) )
-        self.unused = [elem for elem in links.elements if not elem.locked]
+        # simplify the blocks by editing the links so that each track will end up linked to at most one hcal
+        # then recalculate the blocks
+        splitblocks=dict() 
+        
+        for block in self._sorted_block_keys(): #put big interesting blocks first
+            #print "block: ", len(self.blocks[block]),  self.blocks[block].short_name();
+            newblocks=self.simplify_blocks(self.blocks[block],event.history_nodes)
+            if newblocks != None:
+                splitblocks.update( newblocks)      
+        if len(splitblocks):
+            self.blocks.update(splitblocks)
+        
+            
+        #reconstruct each of the resulting blocks        
+        for b in self._sorted_block_keys():  #put big interesting blocks first
+            block=self.blocks[b]
+            if block.is_active: # when blocks are split the original gets deactivated                
+                #ALICE debugging
+                #if len(block.element_uniqueids)<6:
+                #    continue
+                self.reconstruct_block(block)                
+                self.unused.extend( [id for id in block.element_uniqueids if not self.locked[id]])
+                
+        #check if anything is unused
         if len(self.unused):
-            pass
             self.log.warning(str(self.unused))
         self.log.info("Particles:")
-        self.log.info(str(self))
+        self.log.info(str(self))        
             
-    def simplify_group(self, group):
-        # for each track, keeping only the closest hcal link
-        simplified = False
-        assert(len(group)!=0)
-        if len(group)==1:
-            return False
-        tracks = [elem for elem in group if elem.layer=='tracker']
-        for track in tracks:
-            first_hcal = True
-            to_unlink = []
-            for linked in track.linked:
-                if linked.layer == 'hcal_in':
-                    if first_hcal:
-                        first_hcal = False
-                    else:
-                        to_unlink.append(linked)
-            for linked in to_unlink:
-                self.links.unlink(track, linked)
-                simplified = True
-        # remove all ecal-hcal links. ecal linked to hcal give rise to a photon anyway.
-        ecals = [elem for elem in group if elem.layer=='ecal_in']
-        for ecal in ecals:
-            to_unlink = []
-            for linked in ecal.linked:
-                if linked.layer == 'hcal_in':
-                    to_unlink.append(linked)
-            for linked in to_unlink:
-                self.links.unlink(ecal, linked)
-                simplified = True
-        return simplified
-            
-    def reconstruct_group(self, group):
-        particles = []
+      
+    def _sorted_block_keys(self) :
+        #sort blocks (1) by number of elements (2) by mix of ecal, hcal , tracks (the shortname will look like "H1T2" for a block
+        #with one cluster and two tracks)        
+        return sorted(self.blocks.keys(), key=lambda k: (len(self.blocks[k].element_uniqueids), self.blocks[k].short_name()),reverse =True)
         
-        #debugging
+            
+    def simplify_blocks(self, block,history_nodes=None):
+        
+        ''' Block: a block which contains list of element ids and set of edges that connect them
+            history_nodes: optional dictionary of Nodes with element identifiers in each node
+        
+        returns None or a dictionary of new split blocks
+            
+        The goal is to remove, if needed, some links from the block so that each track links to 
+        at most one hcal within a block. In some cases this may separate a block into smaller
+        blocks (splitblocks). The BlockSplitter is used to return the new smaller blocks.
+         If history_nodes are provided then the history will be updated. Split blocks will 
+         have the tracks and cluster elements as parents, and also the original block as a parent
+        '''
+        
+        ids=block.element_uniqueids
+        
+        
+        if len(ids)<=1 :    #no links to remove
+            return  None    
+        
+        # work out any links that need to be removed    
+        #   - for tracks unink all hcals except the closest hcal
+        #   - for ecals unlink hcals
+        to_unlink = []        
+        for id in ids :
+            if Identifier.is_track(id):
+                linked = block.linked_edges(id,"hcal_track") # NB already sorted from small to large distance
+                if linked!=None and len(linked)>1 :
+                    first_hcal = True
+                    for elem in linked:
+                        if first_hcal:
+                            first_dist=elem.distance
+                            first_hcal = False
+                        else:
+                            if (elem.distance==first_dist):
+                                pass
+                            to_unlink.append(elem)
+            elif Identifier.is_ecal(id):
+                # this is now handled  elsewhere and so could be removed
+                # remove all ecal-hcal links. ecal linked to hcal give rise to a photon anyway.
+                linked = block.linked_edges(id,"ecal_hcal")
+                to_unlink.extend(linked)
+        
+        #if there is something to unlink then use the BlockSplitter        
+        splitblocks=None        
+        if len(to_unlink):
+            splitblocks= BlockSplitter(block,to_unlink,history_nodes).blocks
+        
+        return splitblocks
+            
+    def reconstruct_block(self, block):
+        ''' see class description for summary of reconstruction approach
+        '''
+        particles = dict()
+        ids = block.element_uniqueids
+        self.locked = dict()
+        for id in ids:
+            self.locked[id] = False
+        
         self.debugprint = False
+        if (self.debugprint  and len(block.element_uniqueids)> 4):
+            print  block
+            
        
-        
-        #impose additional sorting on elements to allow cross cehcking between two methods
-        #most likely it is not the optimal sorting
-        group.sort( key = lambda  x: ( Identifier.type_short_code(x.uniqueid) ,(-x.energy)))
-        if (self.debugprint  and len(group)> 5) :
-            print  "Group: ", len(group) , " **** " #, group
-            print self.links
-           
-        if len(group)==1: #TODO WARNING!!! LOTS OF MISSING CASES
-            elem = group[0]
-            layer = elem.layer
-            if layer == 'ecal_in' or layer == 'hcal_in':
-                particles.append(self.reconstruct_cluster(elem, layer))
-            elif layer == 'tracker':
-                particles.append(self.reconstruct_track(elem))
-            elem.locked = True
-        else:
-            hcals = [elem for elem in group if elem.layer=='hcal_in']
-            for hcal in hcals:
-                particles.extend(self.reconstruct_hcal(hcal))
-            tracks = [elem for elem in group if elem.layer=='tracker'
-                      and not elem.locked]
-          
-            for track in tracks:
+        if len(ids) == 1: #TODO WARNING!!! LOTS OF MISSING CASES
+            id = ids[0]
+            
+            if Identifier.is_ecal(id):
+                self.insert_particle(block, self.reconstruct_cluster(block.pfevent.ecal_clusters[id],"ecal_in"))
+                
+            elif Identifier.is_hcal(id):
+                self.insert_particle(block, self.reconstruct_cluster(block.pfevent.hcal_clusters[id],"hcal_in"))
+                
+            elif Identifier.is_track(id):
+                self.insert_particle(block, self.reconstruct_track(block.pfevent.tracks[id]))
+                # ask Colin about energy balance - what happened to the associated clusters that one would expect?
+        else: #TODO
+            for id in ids :
+                if Identifier.is_hcal(id):
+                    self.reconstruct_hcal(block,id)
+                    
+            for id in ids :
+                if Identifier.is_track(id) and not self.locked[id]:
                 # unused tracks, so not linked to HCAL
                 # reconstructing charged hadrons.
                 # ELECTRONS TO BE DEALT WITH.
-                particles.append(self.reconstruct_track(track))
-                for elem in track.linked:
+                    self.insert_particle(block, self.reconstruct_track(block.pfevent.tracks[id]))
+                    
                     # tracks possibly linked to ecal->locking cluster
-                    assert(elem.layer == 'ecal_in')
-                    elem.locked = True
+                    for idlink in block.linked_ids(id,"ecal_track"):
+                        #ask colin what happened to possible photons here:
+                        self.locked[idlink] = True
+                        #TODO add in extra photonsbut decide where they should go?
+                        
+                        
             # #TODO deal with ecal-ecal
             # ecals = [elem for elem in group if elem.layer=='ecal_in'
             #          and not elem.locked]
@@ -126,7 +224,41 @@ class PFReconstructor(object):
             #     self.log.warning( 'DEAL WITH ELECTRONS!' ) 
             #     particles.append(self.reconstruct_cluster(ecal, 'ecal_in'))
             #TODO deal with track-ecal
-        return particles 
+          
+    
+       
+    def insert_particle(self, block, newparticle):
+            ''' The new particle will be inserted into the history_nodes (if present).
+                A new node for the particle will be created if needed.
+                It will have as its parents the block and all the elements of the block.
+                '''        
+            #Note that although it may be possible to specify more closely that the particle comes from
+            #some parts of the block, there are frequently ambiguities and so for now the particle is
+            #linked to everything in the block
+            if (newparticle) :
+                newid = newparticle.uniqueid
+                self.particles[newid] = newparticle            
+                
+                #check if history nodes exists
+                if (self.history_nodes == None):
+                    return
+                
+                #find the node for the block        
+                blocknode = self.history_nodes[block.uniqueid]
+                
+                #find or make a node for the particle            
+                if newid  in self.history_nodes :
+                    pnode = self.history_nodes[newid]
+                else :
+                    pnode = Node(newid)
+                    self.history_nodes[newid] = pnode
+                
+                #link particle to the block            
+                blocknode.add_child(pnode)
+                #link particle to block elements
+                for element_id in block.element_uniqueids:
+                    self.history_nodes[element_id].add_child(pnode)    
+    
 
     def neutral_hadron_energy_resolution(self, hcal):
         '''WARNING CMS SPECIFIC! 
@@ -147,47 +279,58 @@ class PFReconstructor(object):
         '''
         
         return 1. + math.exp(-cluster.energy/100.)
+    
+      
+        
+    def reconstruct_hcal(self, block, hcalid):
+        '''
+           block: element ids and edges 
+           hcalid: id of the hcal being processed her
+        
+           has hcal and has a track
+                -> add up all connected tracks, turn each track into a charged hadron
+                -> add up all ecal energies
+                -> if track energies is greater than hcal energy then turn the missing energies into an ecal (photon)
+                      NB this links the photon to the hcal rather than the ecals
+                -> if track energies are less than hcal then make a neutral hadron with rest of hcal energy and turn all ecals into photons
+              has hcal but no track (nb by design there will be no attached ecals because hcal ecal links have been removed)
+                -> make a neutral hadron
+              has hcals
+                -> each hcal is treated using rules above
+        '''
         
         
-    def reconstruct_hcal(self, hcal):      
-        particles = []
+        # hcal used to make ecal_in has a couple of possible issues
         tracks = []
         ecals = []
-        for elem in hcal.linked:
-            assert(elem.layer!='ecal_in') # should have killed ecal-hcal links
-            if elem.layer == 'hcal_in':
-                continue
-            elif elem.layer == 'tracker':
-                tracks.append(elem)
-                for ecal in elem.linked:
-                    if ecal.layer!='ecal_in':
-                        continue
-                    if ecal.locked:
-                        continue
-                    ecals.append(ecal)
-                    ecal.locked = True
-                # ecals.extend([te for te in elem.linked if te.layer=='ecal_in'])
-                # hcal should be the only remaining linked hcal cluster (closest one)
-                thcals = [th for th in elem.linked if th.layer=='hcal_in']
-                assert(thcals[0]==hcal)
+        hcal =block.pfevent.hcal_clusters[hcalid]
         
+        assert(len(block.linked_ids(hcalid, "hcal_hcal"))==0  )
+        trackids =    block.sort_distance_energy(hcalid, block.linked_ids(hcalid, "hcal_track") )
+        for trackid in  trackids:
+            tracks.append(block.pfevent.tracks[trackid])
+            for ecalid in block.linked_ids(trackid, "ecal_track"):
+                # the ecals get all grouped together for all tracks in the block
+                # Maybe we want to link ecals to their closest track etc?
+                # this might help with history work
+                # ask colin.
+                if not self.locked[ecalid]:
+                    ecals.append(block.pfevent.ecal_clusters[ecalid])
+                    self.locked[ecalid]  = True
+                # hcal should be the only remaining linked hcal cluster (closest one)
+                #thcals = [th for th in elem.linked if th.layer=='hcal_in']
+                #assert(thcals[0]==hcal)
         self.log.info( hcal )
         self.log.info( '\tT {tracks}'.format(tracks=tracks) )
         self.log.info( '\tE {ecals}'.format(ecals=ecals) )
         hcal_energy = hcal.energy
-        
-          
-        #print tracks
-        #want tracks to be in order of distance and then in order of high energy to low energy
-        #this is an arbitrary decision but allow consistency for method comparisons consistent 
-        if len(tracks)>1:
-            tracks.sort( key = lambda  x: ( self.links.info(x,hcal) ,(-x.energy)))
-            
-        if len(tracks):    
+        if len(tracks):
             ecal_energy = sum(ecal.energy for ecal in ecals)
             track_energy = sum(track.energy for track in tracks)
             for track in tracks:
-                particles.append(self.reconstruct_track(track))
+                #make a charged hadron
+                self.insert_particle(block, self.reconstruct_track( track))
+                
             delta_e_rel = (hcal_energy + ecal_energy) / track_energy - 1.
             # WARNING
             # calo_eres = self.detector.elements['hcal'].energy_resolution(track_energy)
@@ -195,39 +338,47 @@ class PFReconstructor(object):
             self.log.info( 'dE/p, res = {derel}, {res} '.format(
                 derel = delta_e_rel,
                 res = calo_eres ))
-            if delta_e_rel > self.nsigma_hcal(hcal) * calo_eres:
-                excess = delta_e_rel * track_energy
-                self.log.info( 'excess = {excess:5.2f}, ecal_E = {ecal_e:5.2f}, diff = {diff:5.2f}'.format(
-                    excess=excess, ecal_e = ecal_energy, diff=excess-ecal_energy))
-                if excess <= ecal_energy:
-                    particles.append(self.reconstruct_cluster(hcal, 'ecal_in',
-                                                              excess))
-                else:
-                    particle = self.reconstruct_cluster(hcal, 'hcal_in',
-                                                        excess-ecal_energy)
-                    if particle:
-                        particles.append(particle)
+            if delta_e_rel > self.nsigma_hcal(hcal) * calo_eres: # approx means hcal energy + ecal energies > track energies
+                
+                excess = delta_e_rel * track_energy # energy in excess of track energies
+                #print( 'excess = {excess:5.2f}, ecal_E = {ecal_e:5.2f}, diff = {diff:5.2f}'.format(
+                #    excess=excess, ecal_e = ecal_energy, diff=excess-ecal_energy))
+                if excess <= ecal_energy: # approx means hcal energy > track energies 
+                    # Make a photon from the ecal energy
+                    # We make only one photon using only the combined ecal energies
+                    self.insert_particle(block, self.reconstruct_cluster(hcal, 'ecal_in',excess))
+                    
+                else: # approx means that hcal energy>track energies so we must have a neutral hadron
+                    #excess-ecal_energy is approximately hcal energy  - track energies
+                    self.insert_particle(block, self.reconstruct_cluster(hcal, 'hcal_in',
+                                                        excess-ecal_energy))
                     if ecal_energy:
-                        particles.append(self.reconstruct_cluster(hcal, 'ecal_in',
+                        #make a photon from the remaining ecal energies
+                        #again history is confusingbecause hcal is used to provide direction
+                        #be better to make several smaller photons one per ecal?
+                        self.insert_particle(block, self.reconstruct_cluster(hcal, 'ecal_in',
                                                                   ecal_energy))
 
-        else:
-            # hcal cluster linked to other hcal clusters 
-            for elem in hcal.linked:
-                assert(elem.layer=='hcal_in')
-                # ecal hcal links have been cut
-            particles.append(self.reconstruct_cluster(hcal, 'hcal_in'))
-        hcal.locked = True
-        return particles 
+        else: # case whether there are no tracks make a neutral hadron for each hcal
+              # note that hcal-ecal links have been removed so hcal should only be linked to 
+              # other hcals
+                 
+            self.insert_particle(block,  self.reconstruct_cluster(hcal, 'hcal_in'))
+            
+        self.locked[hcalid] = True
+        
                 
-    def reconstruct_cluster(self, cluster, layer, energy=None, vertex=None):
+    def reconstruct_cluster(self, cluster, layer, energy = None, vertex = None):
+        '''construct a photon if it is an ecal
+           construct a neutral hadron if it is an hcal
+        '''        
         if vertex is None:
             vertex = TVector3()
         pdg_id = None
         if layer=='ecal_in':
-            pdg_id = 22
+            pdg_id = 22 #photon
         elif layer=='hcal_in':
-            pdg_id = 130
+            pdg_id = 130 #K0
         else:
             raise ValueError('layer must be equal to ecal_in or hcal_in')
         assert(pdg_id)
@@ -236,41 +387,42 @@ class PFReconstructor(object):
             energy = cluster.energy
         if energy < mass: 
             return None 
-        
         if (mass==0):
             momentum= energy #avoid sqrt for zero mass
         else:
             momentum = math.sqrt(energy**2 - mass**2)
         p3 = cluster.position.Unit() * momentum
         p4 = TLorentzVector(p3.Px(), p3.Py(), p3.Pz(), energy) #mass is not accurate here
-        particle = Particle(p4, vertex, charge, pdg_id)
+        particle = Reconstructed_Particle(p4, vertex, charge, pdg_id)
         path = StraightLine(p4, vertex)
-        path.points[layer] = cluster.position
+        path.points[layer] = cluster.position #alice: this may be a bit strange because we can make a photon with a path where the point is actually that of the hcal?
+                                            # nb this only is problem if the cluster and the assigned layer are different
         particle.set_path(path)
-        particle.clusters[layer] = cluster
-        cluster.locked = True
+        particle.clusters[layer] = cluster  # not sure about this either when hcal is used to make an ecal cluster?
+        self.locked[cluster.uniqueid] = True #just OK but not nice if hcal used to make ecal.
         if self.debugprint:
-            print  "made particle from cluster ",pdg_id,  cluster, particle     
-        
+            print "made particle from cluster ",pdg_id,  cluster, particle        
         return particle
         
-    def reconstruct_track(self, track, clusters=None):
+    def reconstruct_track(self, track, clusters = None): # cluster argument does not ever seem to be used at present
+        '''construct a charged hadron from the track
+        '''
         vertex = track.path.points['vertex']
         pdg_id = 211 * track.charge
         mass, charge = particle_data[pdg_id]
         p4 = TLorentzVector()
         p4.SetVectM(track.p3, mass)
-        particle = Particle(p4, vertex, charge, pdg_id)
+        particle = Reconstructed_Particle(p4, vertex, charge, pdg_id)
         particle.set_path(track.path)
         particle.clusters = clusters
-        track.locked = True
+        self.locked[track.uniqueid] = True
         if self.debugprint:
-            print "made particle from track ", pdg_id, track, particle
+            print "made particle from track ", pdg_id, track, particle        
         return particle
 
 
     def __str__(self):
-        theStr = ['Particles:']
+        theStr = ['New Rec Particles:']
         theStr.extend( map(str, self.particles))
         theStr.append('Unused:')
         if len(self.unused)==0:
