+++ test/analysis_ee_ZH_cfg.py
@@ -65,7 +65,7 @@ papas = cfg.Analyzer(
     merged_ecals = 'ecal_clusters',
     merged_hcals = 'hcal_clusters',
     tracks = 'tracks',
-    #rec_particles = 'sim_rec_particles', # optional - will only do a simulation reconstruction if a anme is provided
+    #rec_particles = 'sim_rec_particles', # optional - will only do a simulation reconstruction if a name is provided
     rec_particles_no_leptons = 'rec_particles_no_leptons', #only there for verification purposes #TODO make optional
     smeared = 'sim_leptons', 
     history = 'history_nodes',
